+++ main/modules/roboclaw_motor.h
@@ -16,7 +16,7 @@ public:
     void step() override;
     void call(const std::string method_name, const std::vector<ConstExpression_ptr> arguments) override;
 
-    int get_position() const;
+    int64_t get_position() const;
     void power(double value);
     void speed(int value);
 };
